requirements.txt
asgiref==3.6.0
Django==4.1.7
django-admin-honeypot==1.1.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
